Move bisectRight implementation to internal/bisect
